Fix relative path with corrected Angular router failing.
Before This Fix:

Assume the URL of the current page is `/first/second`.
When we use `this.router.navigate(['..'])`, the router
will not work at all. We need to use change the
parameter from `..` to `../../` to navigate up to the
URL `/first`.

The Fix:

We should NOT import the `RouterModule.forRoot([])` in
`app.module` as the routing has already been imported
in `app-routing.module`.

After This Fix:

Assume the URL of the current page is `/first/second`.
When we use `this.router.navigate(['..'])`, the router
will navigate up to `/first` as our expectation.