Show sender on app-post-capture-card. #198
This is implemented with a BAD architecture. The PostCapture wall only display contents from the server,
and the app does NOT save any cache of PostCaptures. The performance will deteriorate for the huge
ovehead from the Internet. This architecture should be redesign with the new (reliable) storage
implementation and the new data model for the exchange between server and app.