Refresh PostCapture whenever navigated and focused
Refresh PostCapture whenever navigated to /home and focus on
PostCaptureTab. When the PostCaptureTab is initially created, it doesn't
need the navigation event (use startwith to force the first emitting),
so landing on home page and go to PostCaptureTab can work correctly, as
well as when the PostCaptureTab component is destoryed and re-created.

Signed-off-by: James Chien <shc261392@gmail.com>