Rebuild Capture - G1 (#474)
* Remove margin at the bottom of PostCaptureTab

Signed-off-by: James Chien <shc261392@gmail.com>

* Add sharable_copy download fallback; add ShareService

Signed-off-by: James Chien <shc261392@gmail.com>

* Improve backButton click event readability

Signed-off-by: James Chien <shc261392@gmail.com>

* Fix broken back navigation from sending-post-capture page

Use first() and untilDestroyed() to make sure the subscription
terminates after the navigation.

Zombie subscription is horrible :zombie:

Signed-off-by: James Chien <shc261392@gmail.com>

* Refresh PostCapture whenever navigated and focused

Refresh PostCapture whenever navigated to /home and focus on
PostCaptureTab. When the PostCaptureTab is initially created, it doesn't
need the navigation event (use startwith to force the first emitting),
so landing on home page and go to PostCaptureTab can work correctly, as
well as when the PostCaptureTab component is destoryed and re-created.

Signed-off-by: James Chien <shc261392@gmail.com>

* Fix race condition caused by multiple subscription

Use single observable to handle next-page data loading. The first batch
of fetching is considered as well, so all relevant race conditions
should be fixed now.

However, the code is a bit messy. Need refactoring later.

Signed-off-by: James Chien <shc261392@gmail.com>