Fetch and update diaBackendAssetId during migration
Previously, diaBackendAssetId is stored in the table of
diaBackendAssetRepository. Since the table is dropped in the current
version, previously registered assets must get their assetId back in
order to be displayed and linked to API data correctly.

The migration process gets all originally owned assets. If any proof in
the ProofRepository lacks diaBackendAssetId and matches one of the
assets with the same oldProofHash, the assetId is stored and the proof
is updated.

Signed-off-by: James Chien <shc261392@gmail.com>