feat(android): Enable large heap size in AndroidManifest.xml
This commit modifies the AndroidManifest.xml file in the
android/app/src/main directory to enable the android:largeHeap="true"
attribute in the <application> element.

Enabling the large heap size allows the application to utilize more memory
resources, which can be advantageous for memory-intensive operations or
applications that require additional memory capacity.