feat(ci): Update GitHub Actions workflow
This commit updates the GitHub Actions workflow in the
`.github/workflows/test.yml` file. The changes include adjusting
the trigger conditions for the workflow. Previously, the workflow
ran on push events for branches `master`, `develop`, `feature-*`,
`fix-*`, and `hotfix-*`. Now, the workflow will also trigger on
pull requests. This modification ensures that the workflow runs
when pull requests are created or updated, providing better
coverage for continuous integration testing.