Ensure push notification confirmation dialog follows ATT confirmation dialog
As the AppsFlyer team mentions, the AppsFlyer SDK should be initialized as soon as possible to collect potential events. Therefore, we initialize the AppsFlyer SDK first. However, before that, we need to display the ATT confirmation dialog. Only after that should we show the push notification confirmation dialog. This sequencing is necessary because iOS does not allow multiple native dialogs to be displayed simultaneously. If we attempt to open a native dialog while another is already on the screen, the system forcefully closes the previous dialog, as mentioned here: [https://pub.dev/packages/app_tracking_transparency#important-notice].