fix(.github/workflows): iOS SDK version issue
Upgrade macOS from version 12 to version 14 to resolve the SDK version
issue.

The app was built with the iOS 16.2 (20C52) SDK. However, starting
April 29, 2024, apps must be built with the iOS 17 SDK or later,
included in Xcode 15 or later. The current macOS-latest is macOS 12,
which uses Xcode 14. We need to update to macOS 14, which uses Xcode
15.
Set up Python version 3.11 for the macOS environment to prevent
installation errors such as
"ModuleNotFoundError: No module named 'distutils'."